mc_pos_control split header
